atfback/requirements.txt
# Core Flask and API
Flask==2.3.3
Flask-CORS==4.0.0
Flask-HTTPAuth==4.8.0
Flask-JWT-Extended==4.4.4

# WSGI Server (production)
gunicorn==21.2.0

# Environment and security
python-dotenv==1.0.0
bcrypt==4.0.1

# OCR and image processing
easyocr==1.7.2
pillow==11.3.0

# Utilities
regex==2023.8.8
requests==2.31.0
numpy==1.24.3
thefuzz==0.22.1
rtree==1.4.1
numpy==1.24.3